New dashboard
Dash is now being used for the dashboard, so requirements got updated and the old dashboard got removed.

I think we should look into using pickled dataframes since it seems to handle both the space requirement as well as the speed to load in Dash. Maybe we can find a work around, but I'm not sure yet. More testing to be done.