Better YoutTubeAccessor as well as Categories
This allows for a lot more and can handle the shitty things that plotly requires. For example, you can't pass values to both x and y for the graphs, you have to pass the indexer. The get_alias allows us to bypass this and get the alias for each thing, but still index them how we want to.