Updates to the accessor, more models and exsperiments
A kinda big bug was fixed from the accessor where it parsed tuples and lists the same when they should not have been. Also looking into more models, and it's also interesting that viewCount is something that doesn't seem to be predictable from basic regression, so we may have to look into more things because finding that it's not easy to predict is quite interesting.