Cleanup, better scripts, boilerplate upgrade
Basically all of the smaller files that used to just do if name main main() now systemexit on main. Better method to do this than just raw calling the function. Removed 2 files that were basically dead and might be revived later. Added another script for a different way to download videos that might be combined with the already existing script.